Flesh out representation of animation and change DrawArea to QGraphicsView.
I've shifted our architecture towards the QGraphicsView framework,
which will handle many of the messy layer-drawing details for us. This
way we don't have to re-render every single layer with each
mouse movement.

This is still largely a prototype. We'll need to continue to shift
the some of the functionality we've implemented into the new system.

Note that I've left DrawArea in the code for the time being so we can pull from its
implementation.

Please review and critique!